ENH: Adds qtButtonDict to hold all qt buttons for questionnaire
The qt button instances will be kept in a dictionary. The key for
the qtButtonDict is the (type, name). The type is currently either
'YesNo' or 'Range'. The name will be the question name parsed from
the input questionnaire XML. The value for 'YesNo' buttons located
in the qtButtonDict will be another dictionary containing the QFrame,
yesRadioButton, and noRadioButton. This may be changed to a class
structure later.