ENH: Creates the XNATDataBaseSession child class of DataBaseSession
The abstract base class DataBaseSession will contains the init method
to get the xml string, create the unreviewed and reviewed lists. It
also has the method to getUnreviewedScan since this will be the same
for all child classes. The XNATDataBaseSession contains it's own method
to get the xml string from XNAT, get the restUrl, parse/create the
unreviewed scans list, and create the columns list from the xml string