ENH: Replaces urllib with a request session for a single XNAT connection
The Python module Requests has a build in session class that allows for
a connection to a server to remain open. Then several requests (such as
Put, Get, and Delete) can be made through the open Requests' session.
This is preferred to the urllib module that creates a new connection to
the server each time it is called. There is a limit to the number of
connections from the same IP Address that can be open with XNAT. With each
review, we need to use both a Get and Put request. Using Requests, these
requests will all reuse the same session.