ENH: LoginCredentials class create popup window for username/password
The LoginCredentials class creates a popup window that asks the user
to enter a username and password. The qt QLineEdit objects for the
username and password start with a default value. Once the user clicks
in the box, the default value is deleted if it still in the box. Then
the user can type new text. The password text is hidden by setting the
echo mode to 2. The LoginCredentials object will return the username
and password when the get methods are called.